Add Lua based value pretty printer
i.e. the the general interfacing and a concrete example for TAP files.
The 'interesting' features of TAP pretty-printing is that some of it is
context dependent and that a lot of lookup tables are used. That means
lookup tables that are part of the current file but also static ones.

All this can be rapidly developer using a Lua module, thus also adjusted
for other formats.